feat(worker): refactor modules and dependencies
