+++ worker/crates/action-universal/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-description = "Re:Earth Flow workflow runner"
-name = "reearth-flow-workflow-runner"
+description = "Re:Earth Flow action universal library"
+name = "reearth-flow-action-universal"
 
 authors.workspace = true
 edition.workspace = true
@@ -12,26 +12,30 @@ version.workspace = true
 
 [dependencies]
 reearth-flow-action-log.workspace = true
-reearth-flow-action-universal.workspace = true
 reearth-flow-action.workspace = true
 reearth-flow-common.workspace = true
 reearth-flow-eval-expr.workspace = true
-reearth-flow-state.workspace = true
 reearth-flow-storage.workspace = true
 reearth-flow-workflow.workspace = true
 
 anyhow.workspace = true
-async-recursion = "1.0.5"
-opentelemetry.workspace = true
-petgraph.workspace = true
+async-trait.workspace = true
+async_zip = {version = "0.0.16", features = ["full"]}
+bytes.workspace = true
+csv = "1.3.0"
+directories.workspace = true
+futures.workspace = true
+rayon.workspace = true
+rhai.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+strum.workspace = true
+strum_macros.workspace = true
 thiserror.workspace = true
+tokio-stream.workspace = true
+tokio-util.workspace = true
 tokio.workspace = true
 tracing-subscriber.workspace = true
 tracing.workspace = true
+typetag.workspace = true
 uuid.workspace = true
-
-[dev-dependencies]
-bytes.workspace = true
-pretty_assertions.workspace = true
