+++ worker/crates/action-universal/src/attribute_manager/runner.rs
@@ -7,8 +7,8 @@ use serde::{Deserialize, Serialize};
 
 use reearth_flow_eval_expr::engine::Engine;
 
-use crate::action::{Action, ActionContext, ActionDataframe, ActionResult, ActionValue};
-use crate::utils::convert_dataframe_to_scope_params;
+use reearth_flow_action::utils::convert_dataframe_to_scope_params;
+use reearth_flow_action::{Action, ActionContext, ActionDataframe, ActionResult, ActionValue};
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
