+++ worker/crates/workflow-runner/src/action_runner.rs
@@ -6,10 +6,10 @@ use reearth_flow_state::State;
 use reearth_flow_workflow::graph::NodeAction;
 use tracing::info_span;
 
-use reearth_flow_action::action::{Action, ActionContext, ActionDataframe};
-#[allow(unused_imports)]
-use reearth_flow_action::prelude::*;
+use reearth_flow_action::{Action, ActionContext, ActionDataframe};
 use reearth_flow_action_log::action_log;
+#[allow(unused_imports)]
+use reearth_flow_action_universal::prelude::*;
 
 pub(crate) struct ActionRunner;
